+++ mirai-core/src/jvmTest/kotlin/net/mamoe/mirai/event/JvmMethodEventsTest.kt
@@ -1,8 +1,8 @@
 /*
- * Copyright 2020 Mamoe Technologies and contributors.
+ * Copyright 2019-2020 Mamoe Technologies and contributors.
  *
- * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
- * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 with Mamoe Exceptions 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AFFERO GENERAL PUBLIC LICENSE version 3 with Mamoe Exceptions license that can be found via the following link.
  *
  * https://github.com/mamoe/mirai/blob/master/LICENSE
  */
@@ -11,12 +11,13 @@
 
 package net.mamoe.mirai.event
 
-import junit.framework.TestCase.assertEquals
 import kotlinx.coroutines.CoroutineScope
-import net.mamoe.mirai.utils.internal.runBlocking
+import kotlinx.coroutines.runBlocking
+import org.jetbrains.annotations.NotNull
 import org.junit.Test
 import java.util.concurrent.atomic.AtomicInteger
 import kotlin.coroutines.EmptyCoroutineContext
+import kotlin.test.assertEquals
 
 
 internal class JvmMethodEventsTest {
@@ -44,13 +45,6 @@ internal class JvmMethodEventsTest {
             @EventHandler
             suspend fun `suspend param Unit`(event: TestEvent) {
                 called.getAndIncrement()
-            }
-
-            @Suppress("unused")
-            @EventHandler
-            suspend fun `suspend param Void`(event: TestEvent): Void? {
-                called.getAndIncrement()
-                return null
             }
 
             @EventHandler
@@ -82,21 +76,28 @@ internal class JvmMethodEventsTest {
 
             @EventHandler
             @Suppress("unused")
+            private fun TestEvent.`test annotations`(@NotNull event: TestEvent): ListeningStatus {
+                called.getAndIncrement()
+                return ListeningStatus.STOPPED
+            }
+
+            @EventHandler
+            @Suppress("unused")
             fun TestEvent.`receiver param LS`(event: TestEvent): ListeningStatus {
                 called.getAndIncrement()
                 return ListeningStatus.STOPPED
             }
         }
 
-//        TestClass().run {
-//            this.registerEvents()
-//
-//            runBlocking {
-//                TestEvent().broadcast()
-//            }
-//
-//            assertEquals(9, this.getCalled())
-//        }
+        TestClass().run {
+            this.registerEvents()
+
+            runBlocking {
+                TestEvent().broadcast()
+            }
+
+            assertEquals(9, this.getCalled())
+        }
     }
 
     @Test